@@ -20,55 +20,55 @@
     "@manuscripts/ext-yjs": "workspace:*",
     "@manuscripts/manuscript-editor": "workspace:*",
     "@manuscripts/manuscript-transform": "workspace:*",
-    "@manuscripts/manuscripts-json-schema": "^1.47.0",
+    "@manuscripts/manuscripts-json-schema": "^1.51.0",
     "@manuscripts/examples-track-editor": "workspace:*",
     "@manuscripts/examples-track-schema": "workspace:*",
     "@manuscripts/examples-track-types": "workspace:*",
     "@manuscripts/style-guide": "workspace:*",
-    "@popperjs/core": "^2.11.5",
-    "history": "^4.10.1",
-    "lib0": "^0.2.42",
+    "@popperjs/core": "^2.11.6",
+    "history": "^5.3.0",
+    "lib0": "^0.2.52",
     "lodash.debounce": "^4.0.8",
-    "mobx": "^6.3.5",
-    "mobx-react": "^7.2.1",
-    "prop-types": "^15.7.2",
-    "prosemirror-commands": "^1.3.0",
+    "mobx": "^6.6.2",
+    "mobx-react": "^7.5.3",
+    "prop-types": "^15.8.1",
+    "prosemirror-commands": "^1.3.1",
     "prosemirror-history": "^1.3.0",
     "prosemirror-keymap": "^1.2.0",
     "prosemirror-model": "^1.18.1",
     "prosemirror-state": "^1.4.1",
     "prosemirror-transform": "^1.7.0",
-    "prosemirror-view": "^1.27.0",
+    "prosemirror-view": "^1.28.1",
     "randomcolor": "^0.6.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-popper": "^2.2.5",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
-    "styled-components": "^5.3.3",
-    "typescript": "4.4.2",
-    "unique-names-generator": "^4.6.0",
-    "uuid": "^8.3.1",
-    "y-prosemirror": "^1.0.14",
-    "y-websocket": "^1.3.18",
-    "yjs": "^13.5.39"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
+    "react-popper": "^2.3.0",
+    "react-router": "^6.4.0",
+    "react-router-dom": "^6.4.0",
+    "styled-components": "^5.3.5",
+    "typescript": "4.8.3",
+    "unique-names-generator": "^4.7.1",
+    "uuid": "^9.0.0",
+    "y-prosemirror": "^1.1.3",
+    "y-websocket": "^1.4.5",
+    "yjs": "^13.5.41"
   },
   "devDependencies": {
-    "@types/history": "^4.7.8",
-    "@types/lodash.debounce": "^4.0.6",
-    "@types/node": "16.7.12",
+    "@types/history": "^4.7.11",
+    "@types/lodash.debounce": "^4.0.7",
+    "@types/node": "18.7.18",
     "@types/randomcolor": "^0.5.6",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router": "^5.1.17",
-    "@types/react-router-dom": "^5.3.2",
-    "@types/styled-components": "^5.1.15",
+    "@types/react": "^18.0.20",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router": "^5.1.19",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.26",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "dotenv": "^10.0.0",
+    "dotenv": "^16.0.2",
     "prosemirror-dev-toolkit": "^1.0.3",
     "typescript": "^4.4.3",
-    "vite": "^3.0.0",
-    "vite-tsconfig-paths": "^3.3.17"
+    "vite": "^3.1.3",
+    "vite-tsconfig-paths": "^3.5.0"
   }
 }
