@@ -19,23 +19,23 @@
     "watch": "rollup -cw"
   },
   "devDependencies": {
-    "@rollup/plugin-alias": "^3.1.8",
-    "@rollup/plugin-commonjs": "^21.0.1",
+    "@rollup/plugin-alias": "^3.1.9",
+    "@rollup/plugin-commonjs": "^22.0.2",
     "@types/bcrypt": "^5.0.0",
     "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/express-serve-static-core": "^4.17.24",
-    "@types/ioredis": "^4.28.1",
-    "@types/jsonwebtoken": "^8.5.5",
+    "@types/express": "^4.17.14",
+    "@types/express-serve-static-core": "^4.17.31",
+    "@types/ioredis": "^4.28.10",
+    "@types/jsonwebtoken": "^8.5.9",
     "@types/morgan": "^1.9.3",
-    "@types/node": "^16.11.6",
-    "@types/ws": "^8.2.0",
-    "dotenv": "^10.0.0",
-    "nodemon": "^2.0.14",
+    "@types/node": "^18.7.18",
+    "@types/ws": "^8.5.3",
+    "dotenv": "^16.0.2",
+    "nodemon": "^2.0.20",
     "rimraf": "^3.0.2",
-    "rollup": "^2.58.3",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "typescript": "^4.4.4"
+    "rollup": "^2.79.0",
+    "rollup-plugin-typescript2": "^0.34.0",
+    "typescript": "^4.8.3"
   },
   "dependencies": {
     "@manuscripts/examples-track-db": "workspace:*",
@@ -43,17 +43,17 @@
     "@manuscripts/manuscript-transform": "workspace:*",
     "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
-    "express": "~4.17.1",
-    "ioredis": "^4.28.1",
-    "joi": "^17.4.2",
-    "jsdom": "^16.4.0",
+    "express": "~4.18.1",
+    "ioredis": "^5.2.3",
+    "joi": "^17.6.0",
+    "jsdom": "^20.0.0",
     "jsonwebtoken": "^8.5.1",
-    "lib0": "^0.2.42",
+    "lib0": "^0.2.52",
     "morgan": "^1.10.0",
-    "winston": "^3.3.3",
-    "ws": "^8.2.3",
-    "y-prosemirror": "^1.0.14",
+    "winston": "^3.8.2",
+    "ws": "^8.8.1",
+    "y-prosemirror": "^1.1.3",
     "y-protocols": "^1.0.5",
-    "yjs": "^13.5.39"
+    "yjs": "^13.5.41"
   }
 }
