@@ -23,13 +23,13 @@
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.0",
-    "@types/node": "^16.11.6",
+    "@types/node": "^18.7.18",
     "bcrypt": "^5.0.1",
-    "prisma": "^4.0.0",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.4.4"
+    "prisma": "^4.3.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.3"
   },
   "dependencies": {
-    "@prisma/client": "^4.0.0"
+    "@prisma/client": "^4.3.1"
   }
 }
