@@ -23,30 +23,30 @@
     "lint:fix": "eslint --fix --ext .js,.ts, ./src"
   },
   "devDependencies": {
-    "@rollup/plugin-alias": "^3.1.8",
-    "@rollup/plugin-commonjs": "^20.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.6",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@zerollup/ts-transform-paths": "^1.7.18",
-    "rollup": "^2.58.3",
-    "rollup-plugin-typescript2": "^0.30.0",
+    "@rollup/plugin-alias": "^3.1.9",
+    "@rollup/plugin-commonjs": "^22.0.2",
+    "@rollup/plugin-node-resolve": "^14.1.0",
+    "@types/react": "^18.0.20",
+    "@types/react-dom": "^18.0.6",
+    "rollup": "^2.79.0",
+    "rollup-plugin-typescript2": "^0.34.0",
     "tslib": "^2.4.0",
-    "ttypescript": "^1.5.12",
-    "typescript": "^4.4.4"
+    "ttypescript": "^1.5.13",
+    "typescript": "^4.8.3",
+    "typescript-transform-paths": "^3.3.1"
   },
   "peerDependencies": {
     "react": ">=17.0.0",
     "react-dom": ">=17.0.0"
   },
   "dependencies": {
-    "@manuscripts/track-changes-plugin": "workspace:*",
     "@manuscripts/examples-track-schema": "workspace:*",
     "@manuscripts/examples-track-types": "workspace:*",
-    "@popperjs/core": "^2.11.5",
+    "@manuscripts/track-changes-plugin": "workspace:*",
+    "@popperjs/core": "^2.11.6",
     "prosemirror-model": "^1.18.1",
     "prosemirror-state": "^1.4.1",
     "prosemirror-transform": "^1.7.0",
-    "prosemirror-view": "^1.27.0"
+    "prosemirror-view": "^1.28.0"
   }
 }
