@@ -24,19 +24,19 @@
   },
   "devDependencies": {
     "@manuscripts/manuscript-editor": "workspace:*",
-    "@rollup/plugin-commonjs": "^20.0.0",
-    "@types/react": "^17.0.20",
+    "@rollup/plugin-commonjs": "^22.0.2",
+    "@types/react": "^18.0.20",
     "@types/uuid": "^8.3.4",
-    "prosemirror-commands": "^1.3.0",
+    "prosemirror-commands": "^1.3.1",
     "prosemirror-keymap": "^1.2.0",
     "prosemirror-model": "^1.18.1",
     "prosemirror-state": "^1.4.1",
-    "prosemirror-transform": "^1.6.0",
-    "prosemirror-view": "^1.27.0",
-    "react": "^17.0.2",
-    "rollup": "^2.58.3",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "typescript": "^4.4.4"
+    "prosemirror-transform": "^1.7.0",
+    "prosemirror-view": "^1.28.1",
+    "react": "^18.2.0",
+    "rollup": "^2.79.0",
+    "rollup-plugin-typescript2": "^0.34.0",
+    "typescript": "^4.8.3"
   },
   "peerDependencies": {
     "@manuscripts/manuscript-editor": "workspace:*",
@@ -46,7 +46,7 @@
     "react": ">=17.0.0"
   },
   "dependencies": {
-    "debug": "^4.1.1",
-    "uuid": "^8.3.2"
+    "debug": "^4.3.4",
+    "uuid": "^9.0.0"
   }
 }
