@@ -25,19 +25,19 @@
   },
   "devDependencies": {
     "@manuscripts/examples-track-editor": "workspace:*",
-    "@rollup/plugin-commonjs": "^20.0.0",
+    "@rollup/plugin-commonjs": "^22.0.2",
     "@types/katex": "^0.14.0",
-    "@types/react": "^17.0.20",
-    "prosemirror-commands": "^1.3.0",
+    "@types/react": "^18.0.20",
+    "prosemirror-commands": "^1.3.1",
     "prosemirror-keymap": "^1.2.0",
     "prosemirror-model": "^1.18.1",
     "prosemirror-state": "^1.4.1",
-    "prosemirror-transform": "^1.6.0",
-    "prosemirror-view": "^1.27.0",
-    "react": "^17.0.2",
-    "rollup": "^2.58.3",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "typescript": "^4.4.4"
+    "prosemirror-transform": "^1.7.0",
+    "prosemirror-view": "^1.28.1",
+    "react": "^18.2.0",
+    "rollup": "^2.79.0",
+    "rollup-plugin-typescript2": "^0.34.0",
+    "typescript": "^4.8.3"
   },
   "peerDependencies": {
     "@manuscripts/examples-track-editor": "workspace:*",
@@ -47,8 +47,8 @@
     "react": ">=17.0.0"
   },
   "dependencies": {
-    "@codemirror/view": "^6.2.0",
+    "@codemirror/view": "^6.2.4",
     "@codemirror/state": "^6.1.1",
-    "katex": "^0.16.0"
+    "katex": "^0.16.2"
   }
 }
