@@ -26,17 +26,17 @@
   },
   "devDependencies": {
     "@manuscripts/examples-track-editor": "workspace:*",
-    "@rollup/plugin-commonjs": "^22.0.1",
-    "prosemirror-commands": "^1.3.0",
+    "@rollup/plugin-commonjs": "^22.0.2",
+    "prosemirror-commands": "^1.3.1",
     "prosemirror-example-setup": "^1.2.1",
     "prosemirror-keymap": "^1.2.0",
     "prosemirror-model": "^1.18.1",
     "prosemirror-state": "^1.4.1",
-    "prosemirror-transform": "^1.6.0",
-    "prosemirror-view": "^1.27.0",
-    "rollup": "^2.77.0",
-    "rollup-plugin-typescript2": "^0.32.1",
-    "typescript": "^4.7.4"
+    "prosemirror-transform": "^1.7.0",
+    "prosemirror-view": "^1.28.1",
+    "rollup": "^2.79.0",
+    "rollup-plugin-typescript2": "^0.34.0",
+    "typescript": "^4.8.3"
   },
   "peerDependencies": {
     "@manuscripts/examples-track-editor": "workspace:*",
