@@ -25,19 +25,19 @@
   },
   "devDependencies": {
     "@manuscripts/manuscript-editor": "workspace:*",
-    "@rollup/plugin-commonjs": "^20.0.0",
+    "@rollup/plugin-commonjs": "^22.0.2",
     "@types/randomcolor": "^0.5.6",
-    "@types/react": "^17.0.20",
-    "prosemirror-commands": "^1.3.0",
+    "@types/react": "^18.0.20",
+    "prosemirror-commands": "^1.3.1",
     "prosemirror-keymap": "^1.2.0",
     "prosemirror-model": "^1.18.1",
     "prosemirror-state": "^1.4.1",
-    "prosemirror-transform": "^1.6.0",
-    "prosemirror-view": "^1.27.0",
-    "react": "^17.0.2",
-    "rollup": "^2.58.3",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "typescript": "^4.4.4"
+    "prosemirror-transform": "^1.7.0",
+    "prosemirror-view": "^1.28.1",
+    "react": "^18.2.0",
+    "rollup": "^2.79.0",
+    "rollup-plugin-typescript2": "^0.34.0",
+    "typescript": "^4.8.3"
   },
   "peerDependencies": {
     "@manuscripts/manuscript-editor": "workspace:*",
@@ -49,13 +49,13 @@
   "dependencies": {
     "@manuscripts/examples-track-types": "workspace:*",
     "@manuscripts/track-changes-plugin": "workspace:*",
-    "debug": "^4.1.1",
-    "lib0": "^0.2.42",
+    "debug": "^4.3.4",
+    "lib0": "^0.2.52",
     "randomcolor": "^0.6.2",
-    "unique-names-generator": "^4.6.0",
-    "y-prosemirror": "^1.0.14",
+    "unique-names-generator": "^4.7.1",
+    "y-prosemirror": "^1.1.3",
     "y-protocols": "^1.0.5",
-    "y-websocket": "^1.3.18",
-    "yjs": "^13.5.39"
+    "y-websocket": "^1.4.5",
+    "yjs": "^13.5.41"
   }
 }
