@@ -24,12 +24,12 @@
     "lint:fix": "eslint --fix --ext .js,.ts, ./src"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^20.0.0",
+    "@rollup/plugin-commonjs": "^22.0.2",
     "prosemirror-model": "^1.18.1",
-    "prosemirror-schema-list": "^1.2.0",
-    "rollup": "^2.58.3",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "typescript": "^4.4.4"
+    "prosemirror-schema-list": "^1.2.2",
+    "rollup": "^2.79.0",
+    "rollup-plugin-typescript2": "^0.34.0",
+    "typescript": "^4.8.3"
   },
   "dependencies": {
     "prosemirror-model": "^1.18.1",
