@@ -17,7 +17,7 @@ import { generateUser, yjsExtension, YjsUser } from '@manuscripts/ext-yjs'
 import { YJS_WS_URL } from 'config'
 import useTrackOptions from 'hooks/useTrackOptions'
 import React, { useEffect, useState } from 'react'
-import { useHistory, useParams } from 'react-router-dom'
+import { useNavigate, useParams } from 'react-router-dom'
 import { stores } from 'stores'
 import styled from 'styled-components'
 import { yDocToProsemirrorJSON } from 'y-prosemirror'
@@ -35,7 +35,7 @@ export function ManuscriptsPage() {
     documentFlows,
     documentStore: { createDocument },
   } = stores
-  const history = useHistory()
+  const navigateTo = useNavigate()
   const routeParams = useParams<{ documentId: string }>()
   const { options, setOptions } = useTrackOptions('manuscript-track-options', {
     documentId: routeParams.documentId,
@@ -57,7 +57,7 @@ export function ManuscriptsPage() {
         // The postgres data is discarded since the Yjs updates are only persisted to Redis for now
       }
     }
-    initialData && history.push(documentId)
+    initialData && navigateTo(documentId)
     const yDoc = new Doc({ gc: false })
     const provider = new WebsocketProvider(YJS_WS_URL, documentId, yDoc)
     provider.on('synced', () => {
